Implement menu, layout, customNodes and validation
